day42/streamsocial-kstream-processor/requirements.txt
kafka-python==2.3.0
six==1.16.0
confluent-kafka==2.3.0
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
python-socketio==5.11.2
requests==2.31.0
prometheus-client==0.20.0
pyyaml==6.0.1
jsonschema==4.21.1
python-dateutil==2.9.0
psutil==5.9.8
pytest==8.0.0
